Add files via upload
Here, we will  find the economic interaction between cities. We will find the total number of orders between cities. 
Means, the economic interaction between city pair "sao paulo" and "rio de janeiro" is the sum of number 
of orders which "sao paulo" as the customer city and "rio de janeiro" as the seller city, and the opposite 
("sao paulo" as the seller city and "rio de janeiro" as the customer city). After calculating economic 
interaction for each city pair returns the top-10 cities having the highest interactions in terms of the 
the number of total orders.
Using Data Set: From https://www.kaggle.com/olistbr/brazilian-ecommerce