Add files via upload
Using “global-city-population-estimates.xls” which contains population numbers 
for cities World-wide. The data is given at every 5 years and range from 1950-2030. Using “CITIESOVER-300K” sheet of this file produce the following figures.
a function that takes the file name and year information as the input.
a) Find the top-10 most populated cities in that year. Plot the result in horizontal bar plot as (10
Points)
• Write city and country names inside the bars. Use bigger fontsize for City name.
• Add year information on the lower-right area of the plot. Color should be red.
• Figure size should be 10x6 inches.
• Put a x-label as “population (thousands)” with fontsize 15
b) Plot the population percentages of these top-10 most populated cities in a pie plot. (10 
Points)
• Figure size should be 6x6 inches.
• Fontsizes for labels and percentage texts should be 15