Add files via upload
This problem is based on a subset of the 311 service requests from NYC Open Data.  The following analyses using 311-service-requests.csv file. 
a) Find the day that has the highest number of complaints. Print the result in the following 
format. (5 Points)
'October-24-Thursday'
b) Produce a table for the number complaint type in each day of the week. Do this analysis only 
for the top-10 most frequent complaints. (7.5 Points)